Copy changes from Wpf
