+++ projects/GKv3/GEDKeeper3.Gtk2/GEDKeeper3.Gtk2.csproj
@@ -42,7 +42,7 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Program.cs" />
+    <Compile Include="GKProgram.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
