+++ projects/GKv3/GEDKeeper3.Gtk2/GKProgram.cs
@@ -1,0 +1,68 @@
+﻿using System;
+using System.Reflection;
+using System.Resources;
+using Eto;
+using Eto.Forms;
+using GKCore;
+using GKUI.Platform;
+
+[assembly: AssemblyTitle("GEDKeeper3.Gtk2")]
+[assembly: AssemblyDescription("")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("GEDKeeper")]
+[assembly: AssemblyCopyright("Copyright © 2009-2017 by Sergey V. Zhdanovskih")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+[assembly: AssemblyVersion("3.0.0.0")]
+[assembly: NeutralResourcesLanguage("en")]
+
+namespace GEDKeeper3.Gtk2
+{
+    /// <summary>
+    /// The main startup class of application.
+    /// </summary>
+    public static class GKProgram
+    {
+        private static void LogSysInfo()
+        {
+            try {
+                //#if MONO
+                //Logger.LogWrite("Mono Version: " + SysUtils.GetMonoVersion());
+                //Logger.LogWrite("Desktop Type: " + SysUtils.GetDesktopType().ToString());
+                //#endif
+
+                // There should be no links to the application infrastructure
+                Assembly execAssembly = Assembly.GetExecutingAssembly();
+                Logger.WriteInfo("CLR Version: " + execAssembly.ImageRuntimeVersion);
+                Logger.WriteInfo("GK Version: " + execAssembly.GetName().Version.ToString());
+            } catch {
+                // dummy
+            }
+        }
+
+        [STAThread]
+        public static void Main(string[] args)
+        {
+            //Style.Add<ButtonToolItemHandler>("icons", h => h.Widget.Image.);
+            //Style.Add<ButtonHandler>("icons", h => h.Widget.Image.);
+
+            EtoAppHost.ConfigureBootstrap(false);
+            AppHost.CheckPortable(args);
+            Logger.Init(AppHost.GetLogFilename());
+            LogSysInfo();
+
+            var application = new Application(Platforms.Gtk2);
+
+            AppHost.InitSettings();
+            try {
+                var appHost = (EtoAppHost) AppHost.Instance;
+                appHost.Init(args, false);
+
+                application.Run();
+            } finally {
+                AppHost.DoneSettings();
+            }
+        }
+    }
+}
