+++ projects/GKv3/GEDKeeper3.Mac/GEDKeeper3.Mac.csproj
@@ -46,7 +46,7 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Program.cs" />
+    <Compile Include="GKProgram.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
