--- projects/GKv3/GEDKeeper3.Mac/Program.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Reflection;
-using Eto;
-using Eto.Forms;
-using GKUI;
-
-[assembly: AssemblyTitle("GEDKeeper3.Mac")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("GEDKeeper")]
-[assembly: AssemblyCopyright("Copyright © 2009-2017 by Sergey V. Zhdanovskih")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: AssemblyVersion("3.0.0.0")]
-
-namespace GEDKeeper3.Mac
-{
-    public class Program
-    {
-        [STAThread]
-        public static void Main(string[] args)
-        {
-            new Application(Platforms.Mac).Run(new BaseWinSDI());
-        }
-    }
-}
